--- src/pages/Usuarios/UsuariosForm/UsuariosForm.module.css
@@ -1,7 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    margin: 40px;
-    row-gap: 18px;
-}
