--- src/pages/Usuarios/UsuariosForm/index.js
@@ -1,1 +1,0 @@
-export  {  default  }  from  './UsuariosForm' ;
